添加音频编码器部分模块

+++ include/module/voltage.h
@@ -6,6 +6,6 @@
 namespace voltage
 {
   void setup();
-  float getVCCVoltage();
-  float getBATVoltage();
+  double getVCCVoltage();
+  double getBATVoltage();
 }
